Fixed problem with no response from /api/server endpoint
Problem was with JsonView. He didn't allow to render data from Server model, because his fields didn't have @JsonView annotation

+++ backend/src/main/kotlin/com/wittano/mineserv/repository/UserRepository.kt
@@ -3,6 +3,6 @@ package com.wittano.mineserv.repository
 import com.wittano.mineserv.data.User
 import org.springframework.data.repository.CrudRepository
 
-interface UserRepository : CrudRepository<User, Int> {
+interface UserRepository : CrudRepository<User, Long> {
     fun findByUsername(username: String): User?
 }
